Add Terraform to deploy the application
This deploys the Lambda, creates the SNS topic and sets up all the
required permissions for the Lambda to run and publish to the topic.

It also subscribes the user's email address to the SNS topic. Due to AWS'
rules, they must click the link in the confirmation email before they
get any of the price alerts.

Notably this does not yet include any way of triggering the lambda