Set lambda timeout to 5s
Last night the function failed with a timeout. I think this is probably
due to the API call being slow because not much else happens in the
Lambda. The extra logging I added will help diagnose where the slowdown
is.

Increase the timeout so we can gather data on what's slow