Handle the API being in UTC time
The query parameters for the API and results from it are in UTC, but I'm
in London which has daylight saving time for half the year.

Calculate 'tomorrow' in my timezone, convert it to UTC for the API query
then convert all times in the response back to my timezone.

This means the time ranges I've picked to include summary data for will
represent the same local time all year.