Increase Lambda timeout and don't run it on the hour
The lambda timed out calling the API again last night. It seems the
90th centile response time can be very slow. Increase the lambda timeout
further to 8 seconds and change the schedule to run at 7 minutes past 8.

I suspect that many people / tools call the API on the hour, so my hope
is that by delaying slightly I'll get a more consistent response time.

If this doesn't fix it I'll look at a retry mechanism with a several
minute backoff.